requirements.txt
flask
requests
fastapi
uvicorn
flask-cors
groq
python-dotenv
requests
jinja2
beautifulsoup4
lxml
pydantic
